Multer profil Image / Media Querries / Authorize user

+++ backend/controllers/publication.js
@@ -22,7 +22,7 @@ exports.createPublication = (req, res, next) => {
   }
   console.log("Bonjour");
   console.log(req.body);
-  let post = new publication(req.body.nom, req.body.prenom, req.body.user_id, req.body.content, filename);
+  let post = new publication(req.body.nom, req.body.prenom, req.body.user_id, req.body.content, filename, req.body.user_image);
   let sql = "INSERT INTO postes SET ?";
   let query = connexion.query(sql,post, (err, res) => {
     if(err) throw err;
@@ -49,42 +49,49 @@ exports.getAllPublication = (req,response, next) => {
 // Modification d'une publication
 
 exports.modifyPublication = (req,res, next) => {
-  let sql = `UPDATE postes SET content = '${req.body.content}' WHERE id = ${req.params.id}`;
-  let query = connexion.query(sql, (err, res) => {
-    if(err) throw err;
-    console.log(res);
-    console.log("Publication Content modifié");
-  });
 
-  if(!req.body.image){
-    let sql2 = `UPDATE postes SET file = '${req.protocol}://${req.get('host')}/images/${req.file.filename}' WHERE id = ${req.params.id}`;
-    let query = connexion.query(sql2, (err, res) => {
-    if(err) throw err;
-    console.log(res);
-    console.log("Publication Image modifié");
-    const filename = req.body.filename.split('http://localhost:3000/images/')[1];
-    fs.unlink(`images/${filename}`, (err) => {
+  if(req.body.id === req.body.user_id || req.body.role === "Admin"){
+    let sql = `UPDATE postes SET content = '${req.body.content}' WHERE id = ${req.params.id}`;
+    let query = connexion.query(sql, (err, res) => {
       if(err) throw err;
-      console.log("file unlinked")
-    })
-  });
+      console.log(res);
+      console.log("Publication Content modifié");
+    });
+  
+    if(!req.body.image){
+      let sql2 = `UPDATE postes SET file = '${req.protocol}://${req.get('host')}/images/${req.file.filename}' WHERE id = ${req.params.id}`;
+      let query = connexion.query(sql2, (err, res) => {
+      if(err) throw err;
+      console.log(res);
+      console.log("Publication Image modifié");
+      const filename = req.body.filename.split('http://localhost:3000/images/')[1];
+      fs.unlink(`images/${filename}`, (err) => {
+        if(err) throw err;
+        console.log("file unlinked")
+      })
+    });
+    }
   }
 };
 
 // Supprimer une publication
 
 exports.deletePublication = (req,res, next) =>{
-  let sql = `DELETE FROM postes WHERE id = ${req.params.id /*req.body.post_delete.post_id*/}`
-  let query = connexion.query(sql, (err, res) => {
-    if(err) throw err;
-    console.log(res);
-    console.log("Publication supprimée");
-    const filename = req.body.filename.split('http://localhost:3000/images/')[1];
-    fs.unlink(`images/${filename}`, (err) => {
-      console.log(`images/${filename}`);
-      if(err) throw err;
-      console.log("file unlinked")
-    })
 
-  })
+  if(req.body.id === req.body.user_id || req.body.role === "Admin"){
+    let sql = `DELETE FROM postes WHERE id = ${req.params.id /*req.body.post_delete.post_id*/}`
+    let query = connexion.query(sql, (err, res) => {
+      if(err) throw err;
+      console.log(res);
+      console.log("Publication supprimée");
+      if(req.body.filename != null){
+        const filename = req.body.filename.split('http://localhost:3000/images/')[1];
+        fs.unlink(`images/${filename}`, (err) => {
+          console.log(`images/${filename}`);
+          if(err) throw err;
+          console.log("file unlinked")
+        })
+      }
+    })
+  }
 }
